pyproject.toml
[project]
name = "rhyme-riddle"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "pillow>=12.1.1",
    "requests>=2.32.5",
]
